Refactor code templates: replace CustomUtility.java, add navigation and layout resources
