--- CodeTemplates/CustomUtility.java
@@ -1,22 +1,0 @@
-package ${PACKAGE_NAME};
-
-// Replace ${PACKAGE_NAME} with your package name
-// Replace ${UTILITY_NAME} with your utility class name
-
-public class ${UTILITY_NAME} {
-
-    private ${UTILITY_NAME}() {
-        // Private constructor to prevent instantiation
-    }
-
-    /**
-     * Static utility method.
-     *
-     * @param param Description of the parameter
-     * @return Description of the return value
-     */
-    public static ReturnType utilityMethod(ParamType param) {
-        // Implement your utility functionality here
-        return result;
-    }
-}
