+++ CodeTemplates/java/util/CustomUtility.java
@@ -1,0 +1,46 @@
+/**
+ * Created by Sim Sze Yu
+ * GitHub: https://github.com/szeyu/Awsome-Android-Code-Template
+ * 
+ * This template is part of the WIA2007 Mobile Application Development Course
+ * Bachelor of Computer Science (Artificial Intelligence)
+ * Universiti Malaya
+ * 
+ * Licensed under MIT License
+ * Copyright (c) 2025 Sim Sze Yu
+ * 
+ * General utility class for common operations.
+ * Features:
+ * - Static helper methods
+ * - Common utility functions
+ * - Reusable operations
+ * - Format conversions
+ * 
+ * Usage:
+ * 1. Import utility class
+ * 2. Call static methods as needed
+ * 3. Add custom utility methods
+ */
+
+package ${PACKAGE_NAME};
+
+// Replace ${PACKAGE_NAME} with your package name
+// Replace ${UTILITY_NAME} with your utility class name
+
+public class ${UTILITY_NAME} {
+
+    private ${UTILITY_NAME}() {
+        // Private constructor to prevent instantiation
+    }
+
+    /**
+     * Static utility method.
+     *
+     * @param param Description of the parameter
+     * @return Description of the return value
+     */
+    public static ReturnType utilityMethod(ParamType param) {
+        // Implement your utility functionality here
+        return result;
+    }
+}
