+++ CodeTemplates/res/layout/custom_button_layout.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Created by Sim Sze Yu
+  ~ GitHub: https://github.com/szeyu/Awsome-Android-Code-Template
+  ~
+  ~ This template is part of the WIA2007 Mobile Application Development Course
+  ~ Bachelor of Computer Science (Artificial Intelligence)
+  ~ Universiti Malaya
+  ~
+  ~ Licensed under MIT License
+  ~ Copyright (c) 2025 Sim Sze Yu
+  ~
+  ~ Material Design button collection template.
+  ~ Features:
+  ~ - Different button styles (contained, outlined, text)
+  ~ - Icon support
+  ~ - Custom corner radius
+  ~ - Consistent padding
+  ~
+  ~ Usage:
+  ~ 1. Include buttons as needed
+  ~ 2. Customize styles
+  ~ 3. Add click handlers
+  -->
+<LinearLayout 
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
+    android:padding="@dimen/spacing_medium">
+
+    <!-- Primary Button -->
+    <com.google.android.material.button.MaterialButton
+        style="@style/Widget.Custom.Button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Primary Button" />
+
+    <!-- Outlined Button -->
+    <com.google.android.material.button.MaterialButton
+        style="@style/Widget.Custom.Button.Outlined"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/spacing_medium"
+        android:text="Outlined Button" />
+
+    <!-- Text Button -->
+    <com.google.android.material.button.MaterialButton
+        style="@style/Widget.MaterialComponents.Button.TextButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/spacing_medium"
+        android:text="Text Button"
+        android:textColor="@color/primary" />
+
+    <!-- Icon Button -->
+    <com.google.android.material.button.MaterialButton
+        style="@style/Widget.Custom.Button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/spacing_medium"
+        android:text="Icon Button"
+        app:icon="@drawable/baseline_add_24"
+        app:iconGravity="textStart"
+        app:iconTint="@color/text_light"
+        app:iconPadding="@dimen/spacing_small" />
+
+</LinearLayout>
